+++ pkg/apis/serving/v1alpha1/inference_graph.go
@@ -17,6 +17,7 @@ limitations under the License.
 package v1alpha1
 
 import (
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"knative.dev/pkg/apis"
@@ -300,4 +301,17 @@ type InferenceGraphList struct {
 
 func init() {
 	SchemeBuilder.Register(&InferenceGraph{}, &InferenceGraphList{})
+}
+
+func (ss *InferenceGraphStatus) PropagateRawStatus(
+	deployment *appsv1.Deployment,
+	url *apis.URL) {
+
+	for _, con := range deployment.Status.Conditions {
+		if con.Type == appsv1.DeploymentAvailable {
+			ss.URL = url
+			break
+		}
+	}
+	ss.ObservedGeneration = deployment.Status.ObservedGeneration
 }
