--- pkg/controller/v1alpha1/inferencegraph/reconcilers/ig_raw_kube_reconciler.go
@@ -1,114 +1,0 @@
-/*
-Copyright 2023 The KServe Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package reconcilers
-
-import (
-	"fmt"
-	"github.com/kserve/kserve/pkg/apis/serving/v1beta1"
-	"github.com/kserve/kserve/pkg/controller/v1beta1/inferenceservice/reconcilers/autoscaler"
-	"github.com/kserve/kserve/pkg/controller/v1beta1/inferenceservice/reconcilers/deployment"
-	"github.com/kserve/kserve/pkg/controller/v1beta1/inferenceservice/reconcilers/ingress"
-	"github.com/kserve/kserve/pkg/controller/v1beta1/inferenceservice/reconcilers/service"
-	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	knapis "knative.dev/pkg/apis"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-)
-
-var log = logf.Log.WithName("RawKubeReconciler")
-
-// RawKubeReconciler reconciles the Native K8S Resources
-type RawKubeReconciler struct {
-	client     client.Client
-	scheme     *runtime.Scheme
-	Deployment *deployment.DeploymentReconciler
-	Service    *service.ServiceReconciler
-	Scaler     *autoscaler.AutoscalerReconciler
-	URL        *knapis.URL
-}
-
-// RawKubeReconciler creates raw kubernetes resource reconciler.
-func NewRawKubeReconciler(client client.Client,
-	scheme *runtime.Scheme,
-	componentMeta metav1.ObjectMeta,
-	componentExt *v1beta1.ComponentExtensionSpec,
-	podSpec *corev1.PodSpec) (*RawKubeReconciler, error) {
-
-	as, err := autoscaler.NewAutoscalerReconciler(client, scheme, componentMeta, componentExt)
-	log.Info("autoscaler as well", "as", as)
-	if err != nil {
-		return nil, err
-	}
-
-	url, err := createRawURL(client, componentMeta)
-	log.Info("url generated", "url", url)
-	log.Info("name and namespace", "name, namespace", componentMeta)
-	if err != nil {
-		return nil, err
-	}
-
-	return &RawKubeReconciler{
-		client:     client,
-		scheme:     scheme,
-		Deployment: deployment.NewDeploymentReconciler(client, scheme, componentMeta, componentExt, podSpec),
-		Service:    service.NewServiceReconciler(client, scheme, componentMeta, componentExt, podSpec),
-		Scaler:     as,
-		URL:        url,
-	}, nil
-}
-
-func createRawURL(client client.Client, metadata metav1.ObjectMeta) (*knapis.URL, error) {
-	ingressConfig, err := v1beta1.NewIngressConfig(client)
-	if err != nil {
-		return nil, err
-	}
-
-	url := &knapis.URL{}
-	url.Scheme = "http"
-	url.Host, err = ingress.GenerateDomainName(metadata.Name, metadata, ingressConfig)
-	if err != nil {
-		return nil, fmt.Errorf("failed creating host name: %v", err)
-	}
-
-	return url, nil
-}
-
-// Reconcile ...
-func (r *RawKubeReconciler) Reconcile() (*appsv1.Deployment, error) {
-	//reconcile Deployment
-	deployment, err := r.Deployment.Reconcile()
-	if err != nil {
-		return nil, err
-	}
-
-	//reconcile Service
-	_, err = r.Service.Reconcile()
-	if err != nil {
-		return nil, err
-	}
-	//reconcile Autoscaler
-	_, err = r.Scaler.Reconcile()
-	log.Info("scaler reconciled as well for inference graph")
-	if err != nil {
-		return nil, err
-	}
-
-	return deployment, nil
-}
